added tests for tab parsing

+++ Models/Note.cs
@@ -1,10 +1,10 @@
 public class Note
 {
   public int pitch { get; set; } // the MIDI number of the note (or what it would be if out of range)
-  public float measureStart { get; set; } // the fraction of the measure before the note starts
+  public double measureStart { get; set; } // the fraction of the measure before the note starts
 
   public Note() {}
-  public Note(int pitch, float measureStart) {
+  public Note(int pitch, double measureStart) {
     this.pitch = pitch;
     this.measureStart = measureStart;
   }
